DRY up publisher specs
Related to the changes made in #73

In preparation of #71 (so we only have to update the specs in one
place when solving it)
